refactor: move `stkpush` and `stkquery` to concerns, load variables in a separate method
